Fixed assert_raises; completed negative test suite
assert_raises wouldn't raise an AssertionError if the the wrong kind of
exception was raised in the block being tested, only if no exception was
raised. This seemed counterintuitive, so we now have an
assert_isinstance check inside a more general except block.

Also completed the test suite so now all assertions are covered both
positively and negatively, and all pass in Python 2.5-3.3.